Move 'LockOperations' to management 'IModelsClient'
