+++ clients/imodels-client-authoring/src/base/types/apiEntities/LockErrorInterfaces.ts
@@ -2,33 +2,11 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { IModelsErrorBase } from "@itwin/imodels-client-management";
-
-import { LockLevel } from "./LockInterfaces";
-
-/** Error thrown by Lock update operation. */
-export interface LocksError extends IModelsErrorBase {
-  /** Object ids that are causing the Lock update error. */
-  objectIds?: string[];
-}
-
-/**
- * Error thrown by Lock update operation in case Locks cannot be updated because of conflicts with other Briefcases.
- */
-export interface ConflictingLocksError extends IModelsErrorBase {
-  /** List of locks that are causing the conflicts. */
-  conflictingLocks?: ConflictingLock[];
-}
-
-/** Detailed information about a particular object Lock that is causing the Lock update conflict. */
-export interface ConflictingLock {
-  /** Id of the object that is causing conflict. */
-  objectId: string;
-  /**
-   * The level of conflicting lock. Possible values are {@link LockLevel.Shared}, {@link LockLevel.Exclusive}.
-   * See {@link LockLevel}.
-   */
-  lockLevel: LockLevel;
-  /** An array of Briefcase ids that hold this lock. */
-  briefcaseIds: number[];
-}
+export {
+  /** @deprecated in 6.1. Use `ConflictingLock` from `@itwin/imodels-client-management` instead. */
+  ConflictingLock,
+  /** @deprecated in 6.1. Use `ConflictingLocksError` from `@itwin/imodels-client-management` instead. */
+  ConflictingLocksError,
+  /** @deprecated in 6.1. Use `LocksError` from `@itwin/imodels-client-management` instead. */
+  LocksError,
+} from "@itwin/imodels-client-management";
