+++ clients/imodels-client-management/src/operations/IModelsApiUrlFormatter.ts
@@ -14,6 +14,7 @@ import { GetChangesetGroupListUrlParams } from "./changeset-group/ChangesetGroup
 import { CheckpointParentEntityId } from "./checkpoint/CheckpointOperationParams";
 import { GetFavoriteIModelListUrlParams } from "./favorite-imodel/FavoriteIModelOperationParams";
 import { GetIModelListUrlParams } from "./imodel/IModelOperationParams";
+import { GetLockListUrlParams } from "./lock/LockOperationParams";
 import { GetNamedVersionListUrlParams } from "./named-version/NamedVersionOperationParams";
 import { DownloadThumbnailUrlParams } from "./OperationParamExports";
 import { GetRecentIModelListUrlParams } from "./recent-imodel/RecentIModelOperationParams";
@@ -242,6 +243,15 @@ export class IModelsApiUrlFormatter {
     iModelId: string;
   }): string {
     return `${this.baseUrl}/${params.iModelId}/operations/create`;
+  }
+
+  public getLockListUrl(params: {
+    iModelId: string;
+    urlParams?: GetLockListUrlParams;
+  }): string {
+    return `${this.baseUrl}/${params.iModelId}/locks${this.formQueryString({
+      ...params.urlParams,
+    })}`;
   }
 
   public parseCheckpointUrl(
